refactor: extract dm_repository into its own package (#2894)
Move DM repository, decryption worker, sync state, and NIP-17 message
service into `packages/dm_repository/`. This removes ~5,600 lines of
tests from the main CI shard and gives the DM layer its own CI workflow.

- Extract `IsolateDecryptSigner` interface to `nostr_sdk` (shared by
  `local_key_signer`, `nostr_identity`, and the new package)
- Move 4 source files and 6 test files via `git mv` (preserves history)
- Update 27 app-side imports to `package:dm_repository`
- Register in workspace and add `.github/workflows/dm_repository.yaml`
