+++ mobile/lib/blocs/dm/conversation/conversation_bloc.dart
@@ -6,9 +6,9 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
+import 'package:dm_repository/dm_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:models/models.dart';
-import 'package:openvine/repositories/dm_repository.dart';
 
 part 'conversation_event.dart';
 part 'conversation_state.dart';
