+++ mobile/lib/blocs/dm/conversation_actions/conversation_actions_cubit.dart
@@ -2,8 +2,8 @@
 // ABOUTME: Constructor-injected services, no Riverpod at action time.
 
 import 'package:bloc/bloc.dart';
+import 'package:dm_repository/dm_repository.dart';
 import 'package:equatable/equatable.dart';
-import 'package:openvine/repositories/dm_repository.dart';
 import 'package:openvine/services/content_blocklist_service.dart';
 import 'package:openvine/services/content_moderation_service.dart';
 import 'package:openvine/services/content_reporting_service.dart';
