+++ mobile/lib/blocs/dm/conversation_list/conversation_list_bloc.dart
@@ -5,9 +5,9 @@
 
 import 'package:bloc/bloc.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
+import 'package:dm_repository/dm_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:models/models.dart';
-import 'package:openvine/repositories/dm_repository.dart';
 import 'package:openvine/repositories/follow_repository.dart';
 import 'package:openvine/services/content_blocklist_service.dart';
 import 'package:rxdart/rxdart.dart';
