+++ mobile/packages/dm_repository/lib/dm_repository.dart
@@ -1,0 +1,4 @@
+export 'src/dm_decryption_worker.dart';
+export 'src/dm_repository.dart';
+export 'src/dm_sync_state.dart';
+export 'src/nip17_message_service.dart';
