+++ mobile/packages/dm_repository/pubspec.yaml
@@ -1,0 +1,30 @@
+name: dm_repository
+description: DM repository layer — NIP-17 private messaging, sync, and decryption.
+version: 0.1.0+1
+publish_to: none
+resolution: workspace
+
+environment:
+  sdk: ^3.11.0
+
+dependencies:
+  crypto: ^3.0.7
+  db_client:
+    path: ../db_client
+  flutter:
+    sdk: flutter
+  models:
+    path: ../models
+  nostr_client:
+    path: ../nostr_client
+  nostr_sdk:
+    path: ../nostr_sdk
+  shared_preferences: ^2.5.3
+  unified_logger:
+    path: ../unified_logger
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  mocktail: ^1.0.4
+  very_good_analysis: ^10.2.0
