+++ mobile/packages/dm_repository/test/src/dm_repository_cold_start_test.dart
@@ -5,13 +5,12 @@
 // ABOUTME: docs/plans/2026-04-05-dm-scaling-fix-design.md.
 
 import 'package:db_client/db_client.dart';
+import 'package:dm_repository/dm_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:nostr_client/nostr_client.dart';
 import 'package:nostr_sdk/event.dart';
 import 'package:nostr_sdk/signer/local_nostr_signer.dart';
-import 'package:openvine/repositories/dm_repository.dart';
-import 'package:openvine/services/nip17_message_service.dart';
 
 class _MockNostrClient extends Mock implements NostrClient {}
