+++ mobile/packages/nostr_sdk/lib/signer/isolate_decrypt_signer.dart
@@ -1,0 +1,8 @@
+import 'nostr_signer.dart';
+
+/// Optional capability for signers that can expose a local private key to an
+/// isolate-safe decrypt pipeline.
+abstract interface class IsolateDecryptSigner implements NostrSigner {
+  bool get canDecryptInIsolate;
+  T withPrivateKeyHex<T>(T Function(String hex) operation);
+}
