+++ mobile/pubspec.yaml
@@ -31,6 +31,7 @@ workspace:
   - packages/divine_ui
   - packages/db_client
   - packages/divine_camera
+  - packages/dm_repository
   - packages/divine_video_player
   - packages/funnelcake_api_client
   - packages/hashtag_repository
@@ -104,6 +105,7 @@ dependencies:
     path: packages/divine_camera
   divine_video_player:
     path: packages/divine_video_player
+  dm_repository:
   video_player: ^2.9.1 # Uses native AVPlayer (iOS/macOS) and ExoPlayer (Android)
   visibility_detector: ^0.4.0+2 # Detect widget visibility for proper video control
   image: ^4.1.7 # Image processing for GIF generation
