+++ mobile/test/integration/bug_report_worker_api_test.dart
@@ -3,6 +3,7 @@
 
 import 'dart:io';
 
+import 'package:dm_repository/dm_repository.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
@@ -11,7 +12,6 @@ import 'package:models/models.dart'
 import 'package:nostr_client/nostr_client.dart';
 import 'package:openvine/services/blossom_upload_service.dart';
 import 'package:openvine/services/bug_report_service.dart';
-import 'package:openvine/services/nip17_message_service.dart';
 
 class _MockNIP17MessageService extends Mock implements NIP17MessageService {}
