+++ mobile/test/screens/inbox/conversation/conversation_page_test.dart
@@ -2,11 +2,11 @@
 // ABOUTME: route constants. Ensures ConversationPage provides ConversationBloc
 // ABOUTME: to ConversationView via BlocProvider.
 
+import 'package:dm_repository/dm_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/providers/user_profile_providers.dart';
-import 'package:openvine/repositories/dm_repository.dart';
 import 'package:openvine/screens/inbox/conversation/conversation_page.dart';
 import 'package:openvine/screens/inbox/conversation/conversation_view.dart';
 import 'package:openvine/services/auth_service.dart';
